sellMenuCommand with a pattern to create dynamic List of buttons. Db dump because of incompatible types in several models.

+++ eAdvertisement_bot/Models/DbEntities/AutoBuy.cs
@@ -29,7 +29,7 @@ namespace eAdvertisement_bot.Models.DbEntities
         public int Max_Cpm { get; set; }
 
         [Column("user_id")]
-        public int User_Id { get; set; }
+        public long User_Id { get; set; }
         public User User { get; set; }
 
         // Many To Many relationship lists
